corrected player evaluation computation, so it is more accurate
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/branches/ai@10426 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ trunk/src/karts/controller/fuzzy_ai_controller.hpp
@@ -179,8 +179,8 @@ private:
                                   std::vector<float> parameters );
 
     int   computePlayerEvaluation (unsigned int number_of_players,
-                                   unsigned int player_average_rank,
-                                   unsigned int player_crash_count );
+                                   float        player_average_rank,
+                                   float        player_crash_count );
 
     int   computeCompetitiveness  (unsigned int number_of_players,
                                    int          player_level,
